pyproject.toml
[tool.poetry]
name = "rag-librarian"
version = "0.1.0"
description = "A local-first 'Librarian' service for RAG."
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
pyyaml = "^6.0.1"
qdrant-client = "^1.7.0"
pymupdf = "^1.23.0"
ollama = "^0.1.6"
requests = "^2.31.0"
pdfplumber = "^0.10.3"
jinja2 = "^3.1.3"
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
ruff = "^0.1.14"
mypy = "^1.8.0"
black = "^24.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
